Move 'loginNew' to the actual login page, make Meghan's tweaks
